+++ CHANGELOG.md
@@ -7,6 +7,10 @@ All notable changes to this project will be documented in this file.
 Added new props:
 
 - cursor
+- border-bottom-left-radius
+- border-bottom-right-radius
+- border-top-left-radius
+- border-top-right-radius
 
 ### v0.0.3